+++ http4k-core/src/main/kotlin/org/http4k/events/AutoJsonEvents.kt
@@ -4,6 +4,6 @@ import org.http4k.format.AutoMarshallingJson
 
 object AutoJsonEvents {
     operator fun invoke(json: AutoMarshallingJson, print: (String) -> Unit = ::println) = object : Events {
-        override fun invoke(p1: Event) = print(json.asString(p1))
+        override fun invoke(p1: Event) = print(json.asFormatString(p1))
     }
 }