+++ CHANGELOG.md
@@ -10,7 +10,7 @@ changes with their rationale when appropriate:
 - [http4k-cloudnative] - Fix #418 - Fix separator propagation when adding values to an existing MapEnvironment. H/T @jshiell
 - [http4k-contract] - Add support for securing the API description endpoint. H/T @goodhoko for the PR.
 - [http4k-client-websocket] Added auto-reconnection support on blocking WsClient. H/T @alphaho for the PR.
-- [http4k-format-*] Rename/deprecate `asXYZString(Any)` -> `asString(Any)`
+- [http4k-format-*] Rename/deprecate `asXYZString(Any)` -> `asFormatString(Any)` in all modules
 
 ### v3.247.0
 - [http4k-server-ktornetty] New backend module! H/T @albertlatacz for the contribution!